Refactorización: se han movido los elementos genéricos de CollectionScreen.kt a HeroComposables.kt.
