+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/components/hero/HeroComposables.kt
@@ -2,6 +2,7 @@ package ar.edu.unlam.mobile.scaffold.ui.components.hero
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
@@ -9,10 +10,13 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardColors
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -21,7 +25,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import ar.edu.unlam.mobile.scaffold.domain.hero.DataHero
 
@@ -124,6 +130,64 @@ fun HeroItem(modifier: Modifier = Modifier, hero: DataHero = DataHero()) {
             Text(
                 modifier = Modifier.padding(2.dp),
                 text = hero.name
+            )
+        }
+    }
+}
+
+@Composable
+fun HeroGallery(
+    modifier: Modifier = Modifier,
+    heroList: List<DataHero>,
+    itemSize: Dp = 125.dp,
+    onItemClick: (Int) -> Unit
+) {
+    LazyVerticalGrid(
+        modifier = modifier,
+        columns = GridCells.Adaptive(itemSize),
+        content = {
+            items(heroList.size) {i ->
+                GalleryItem(
+                    dataHero = heroList[i],
+                    onItemClick = onItemClick
+                )
+            }
+        }
+    )
+}
+
+@Composable
+fun GalleryItem(
+    modifier: Modifier = Modifier,
+    dataHero: DataHero = DataHero(),
+    onItemClick : (Int) -> Unit,
+) {
+    TextButton(
+        modifier = modifier,
+        shape = RectangleShape,
+        onClick = { onItemClick(dataHero.id.toInt()) }
+    ) {
+        Box(
+            modifier = Modifier
+                .background(brush = SolidColor(Color.Black), alpha = 0.25f)
+        ) {
+            Column {
+                HeroImage(
+                    modifier = Modifier.padding(4.dp),
+                    url = dataHero.image.url
+                )
+                Text(dataHero.name,
+                    modifier = Modifier
+                        .padding(bottom = 4.dp, start = 4.dp, end = 4.dp)
+                        .align(Alignment.CenterHorizontally),
+                    color = Color.Black,
+                    textAlign = TextAlign.Center
+                )
+            }
+            Text(
+                text = dataHero.id,
+                modifier = Modifier.padding(8.dp),
+                color = Color.Black
             )
         }
     }
