+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/screens/collection/CollectionScreen.kt
@@ -1,33 +1,19 @@
 package ar.edu.unlam.mobile.scaffold.ui.screens.collection
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import ar.edu.unlam.mobile.scaffold.R
-import ar.edu.unlam.mobile.scaffold.domain.hero.DataHero
-import ar.edu.unlam.mobile.scaffold.ui.components.hero.HeroImage
+import ar.edu.unlam.mobile.scaffold.ui.components.hero.HeroGallery
 
 @Composable
 fun CollectionScreen(
@@ -52,64 +38,6 @@ fun CollectionScreen(
                 modifier = Modifier.fillMaxSize(),
                 heroList = viewModel.heroList.collectAsStateWithLifecycle().value,
                 onItemClick = onItemClick
-            )
-        }
-    }
-}
-
-@Composable
-fun HeroGallery(
-    modifier: Modifier = Modifier,
-    heroList: List<DataHero>,
-    itemSize: Dp = 125.dp,
-    onItemClick: (Int) -> Unit
-) {
-    LazyVerticalGrid(
-        modifier = modifier,
-        columns = GridCells.Adaptive(itemSize),
-        content = {
-            items(heroList.size) {i ->
-                GalleryItem(
-                    dataHero = heroList[i],
-                    onItemClick = onItemClick
-                )
-            }
-        }
-    )
-}
-
-@Composable
-fun GalleryItem(
-    modifier: Modifier = Modifier,
-    dataHero: DataHero = DataHero(),
-    onItemClick : (Int) -> Unit,
-) {
-    TextButton(
-        modifier = modifier,
-        shape = RectangleShape,
-        onClick = { onItemClick(dataHero.id.toInt()) }
-    ) {
-        Box(
-            modifier = Modifier
-                .background(brush = SolidColor(Color.Black), alpha = 0.25f)
-        ) {
-            Column {
-                HeroImage(
-                    modifier = Modifier.padding(4.dp),
-                    url = dataHero.image.url
-                )
-                Text(dataHero.name,
-                    modifier = Modifier
-                        .padding(bottom = 4.dp, start = 4.dp, end = 4.dp)
-                        .align(Alignment.CenterHorizontally),
-                    color = Color.Black,
-                    textAlign = TextAlign.Center
-                )
-            }
-            Text(
-                text = dataHero.id,
-                modifier = Modifier.padding(8.dp),
-                color = Color.Black
             )
         }
     }
